# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrustDesk API - Verifiable support platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID (staff only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithRisk"
  /users/{userId}/ban:
    post:
      operationId: banUser
      tags: [staff]
      summary: Ban a user (staff only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # Dashboard (wow endpoints)
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    User:
      type: object
      properties:
        id:
          type: string
        primaryWallet:
          type: string
          nullable: true
        primaryEmail:
          type: string
          nullable: true
        reputationScore:
          type: number
        riskScore:
          type: number
        isStaff:
          type: boolean
        isBanned:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, reputationScore, riskScore, isStaff, isBanned, createdAt]
    UserWithRisk:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            recentRiskEvents:
              type: array
              items:
                $ref: "#/components/schemas/RiskEvent"
          required: [recentRiskEvents]
    RiskEvent:
      type: object
      properties:
        id:
          type: string
        eventType:
          type: string
        scoreDelta:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, eventType, createdAt]
...